+++ changelog.md
@@ -17,6 +17,7 @@ Release 0.13.3 2/7/2022
 
 Release 0.13.2 1/24/2022
 
+- Added support for Unmodified Damage in the GCS and GCA4/5 files for use with GGA import. (@Stevil)
 - Added support for c:[] d:[] p:[] f:[] parsing from notes
 - Added support for upcoming vtt_notes from GCS
 - Added /show FP HP